+++ annotations/docker-annotations/src/main/resources/META-INF/services/io.dekorate.BuildServiceFactory
@@ -1,1 +1,1 @@
-io.dekorate.docker.DockerBuildServiceFactory
+io.dekorate.docker.buildservice.DockerBuildServiceFactory
